+++ packages/delta-protocol-test/src/tests/client-impl.tests.ts
@@ -19,10 +19,13 @@ import { throws } from "assert"
 import { expect } from "chai"
 import { expectError } from "../test-utils/async.js"
 
-import { LionWebClient } from "@lionweb/delta-protocol-client"
+import { LionWebClient, LionWebDeltaProtocolLowLevelClientInstantiator } from "@lionweb/delta-protocol-client"
 import {
+    Command,
+    Event,
     PartitionAddedEvent,
     PartitionDeletedEvent,
+    QueryMessage,
     semanticConsoleLogger,
     SignOffResponse,
     SignOnResponse,
@@ -133,24 +136,24 @@ describe("implementation of LionWeb client", async function() {
                     "cmd-1": {
                         messageKind: "PartitionAdded",
                         newPartition: {
-                            "serializationFormatVersion": "2023.1",
-                            "languages": [
-                                { "key": "TestLanguage", "version": "0" },
-                                { "key": "LionCore-builtins", "version": "2023.1" }
+                            serializationFormatVersion: "2023.1",
+                            languages: [
+                                { key: "TestLanguage", version: "0" },
+                                { key: "LionCore-builtins", version: "2023.1" }
                             ],
-                            "nodes": [
+                            nodes: [
                                 {
-                                    "id": "partition-A",
-                                    "classifier": { "language": "TestLanguage", "version": "0", "key": "TestPartition" },
-                                    "properties": [],
-                                    "containments": [],
-                                    "references": [],
-                                    "annotations": [],
-                                    "parent": null
+                                    id: "partition-A",
+                                    classifier: testLanguageBase.TestPartition.metaPointer(),
+                                    properties: [],
+                                    containments: [],
+                                    references: [],
+                                    annotations: [],
+                                    parent: null
                                 }
                             ]
                         },
-                        sequenceNumber: 0,
+                        sequenceNumber: 1,
                         originCommands: [
                             {
                                 participationId: "participation-a",
@@ -163,7 +166,7 @@ describe("implementation of LionWeb client", async function() {
                         messageKind: "PartitionDeleted",
                         deletedPartition: "partition-A",
                         deletedDescendants: [], // (partition consists of only its root)
-                        sequenceNumber: 1,
+                        sequenceNumber: 2,
                         originCommands: [
                             {
                                 participationId: "participation-a",
@@ -205,6 +208,50 @@ describe("implementation of LionWeb client", async function() {
         lionWebClient.deletePartition(partitionA)
 
         expect(lionWebClient.forest.partitions).to.deep.equal([])
+    })
+
+    it("partition gets added to the client’s forest after receiving a PartitionAdded event", async function() {
+        let receiveMessageOnClientIntercepted: (message: Event | QueryMessage) => void
+        const lowLevelClientInstantiator: LionWebDeltaProtocolLowLevelClientInstantiator = ({ receiveMessageOnClient }) => {
+            receiveMessageOnClientIntercepted = receiveMessageOnClient
+            return Promise.resolve({
+                sendMessage: (_: Command | QueryMessage) => Promise.resolve(),
+                disconnect: () => Promise.resolve()
+            })
+        }
+        const lionWebClient = await LionWebClient.create({
+            repositoryId: "myRepo",
+            clientId: "A",
+            url: "",
+            languageBases: [testLanguageBase],
+            lowLevelClientInstantiator
+        })
+        expect(lionWebClient.forest.partitions.length).to.equal(0)
+
+        expect(receiveMessageOnClientIntercepted!).to.not.equal(undefined)
+        const receiveMessageOnClient = receiveMessageOnClientIntercepted!
+
+        receiveMessageOnClient({
+            messageKind: "PartitionAdded",
+            newPartition: {
+                nodes: [
+                    {
+                        id: "testPartition",
+                        classifier: testLanguageBase.TestPartition.metaPointer(),
+                        properties: [],
+                        containments: [],
+                        references: [],
+                        annotations: [],
+                        parent: null
+                    }
+                ]
+            },
+            sequenceNumber: 1,
+            originCommands: [],
+            additionalInfos: []
+        } as PartitionAddedEvent)
+
+        expect(lionWebClient.forest.partitions.length).to.equal(1)
     })
 
 })