+++ packages/delta-protocol-test/src/tests/statechange-driven-client.tests.ts
@@ -22,12 +22,12 @@ import {
     applyDelta,
     deltaDeserializer,
     DeltaReceiver,
+    deltaSerializer,
     IdMapping,
     INodeBase,
     nodeBaseDeserializer,
     nodeBaseDeserializerWithIdMapping,
-    SerializedDelta,
-    serializeDelta
+    SerializedDelta
 } from "@lionweb/class-core"
 import { LowLevelClient } from "@lionweb/delta-protocol-client"
 import {
@@ -113,6 +113,8 @@ describe("WebSocket-driven client and repository", async function() {
             createWebSocketServer<void, Payload, void, Payload>(port, (_) => undefined, receiveMessageOnServer, repositoryLogger),
             createWSLowLevelClient<Payload, Payload>({ url: wsLocalhostUrl(port), clientId, receiveMessageOnClient }, asLowLevelClientLogger(clientLogger))
         ])
+
+        const serializeDelta = deltaSerializer()
 
         let loading = true
         let commandNumber = 0