+++ packages/json-diff/src/changes/PropertyChange.ts
@@ -1,7 +1,7 @@
-import { LionWebId, LionWebJsonMetaPointer, LionWebJsonNode } from "@lionweb/json"
+import { LionWebId, LionWebJsonMetaPointer } from "@lionweb/json"
 import { JsonContext } from "@lionweb/json-utils"
 import { asMinimalJsonString } from "@lionweb/ts-utils"
-import { Change, Missing } from "./Change.js"
+import { Change } from "./Change.js"
 
 export abstract class PropertyChange extends Change {
     constructor(
@@ -10,7 +10,6 @@ export abstract class PropertyChange extends Change {
         public property: LionWebJsonMetaPointer,
         public oldValue: string | null,
         public newValue: string | null,
-        public missing: Missing = Missing.NotMissing
     ) {
         super(context)
     }
@@ -19,18 +18,33 @@ export abstract class PropertyChange extends Change {
 export class PropertyValueChanged extends PropertyChange {
     readonly changeType = "PropertyValueChanged"
     protected msg = () =>
-        `Node "${this.nodeId} changed value of property "${this.property.key}" from "${this.oldValue}" to "${this.newValue}"`
+        `Node "${this.nodeId} property "${this.property.key}" value changed from "${this.oldValue}" to "${this.newValue}"`
 }
 
-export class PropertyAdded extends Change {
+export class PropertyAdded extends PropertyChange {
     readonly changeType = "PropertyAdded"
     constructor(
-        ctx: JsonContext,
-        public node: LionWebJsonNode,
-        public property: LionWebJsonMetaPointer
+        context: JsonContext,
+        nodeId: LionWebId,
+        property: LionWebJsonMetaPointer,
+        newValue: string | null,
     ) {
-        super(ctx)
+        super(context, nodeId, property, null, newValue)
     }
 
-    protected msg = () => `Node "${this.node.id}" containment added: "${asMinimalJsonString(this.property)}"`
+    protected msg = () => `Node "${this.nodeId}' property added: '${asMinimalJsonString(this.property)}' with value '${this.newValue}'`
+}
+
+export class PropertyDeleted extends PropertyChange {
+    readonly changeType = "PropertyDeleted"
+    constructor(
+        context: JsonContext,
+        nodeId: LionWebId,
+        property: LionWebJsonMetaPointer,
+        oldValue: string | null,
+    ) {
+        super(context, nodeId, property, oldValue, null)
+    }
+
+    protected msg = () => `Node "${this.nodeId}" property deleted: "${asMinimalJsonString(this.property)}"`
 }