+++ packages/cli/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.10.0 — not yet released
 
 * Add a `--sort-connections` flag to the `sort` command — see [`README.md`](README.md) for more information.
+* The `generate-class-core-types` command determines the applicable LionWeb version from the given serialization chunk.
 
 
 ## 0.9.0