+++ packages/json-diff/CHANGELOG.md
@@ -1,8 +1,13 @@
 # Changelog
 
-## 0.10.0
+## 0.10.0 — not yet released
 
-Fixing bug in diffing references.
+* Refactor JSON diffing w.r.t. property changes:
+  * Rename `Missing` to `FeatureMissing`.
+  * Split `PropertyChanged` into three types: `Changed`, `Added`, `Deleted`.
+  * Refactor `LionWebJsonDiff` with above changes.
+* Fix bug in diffing references.
+
 
 ## 0.9.0, 0.8.0, 0.7.2